Renamed a package to internal in the org.eclipse.smarthome.storage.json bundle. See openhab/static-code-analysis#100
