+++ bundles/storage/org.eclipse.smarthome.storage.json/src/main/java/org/eclipse/smarthome/storage/json/internal/ConfigurationDeserializer.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.storage.json;
+package org.eclipse.smarthome.storage.json.internal;
 
 import java.lang.reflect.Type;
 import java.math.BigDecimal;
